refactor: simplify path cache manager and remove unused timeout logic
