--- src/utils/path-cache/index.ts
@@ -1,7 +1,0 @@
-export { PathCacheManager } from './PathCacheManager';
-export type {
-  CacheEntry,
-  PathSearchOptions,
-  PathSearchResult,
-  SearchCacheKey,
-} from './types';
